+++ src/index.js
@@ -1,5 +1,7 @@
 export { loadManifest, validateManifest } from "./config.js";
 export { evaluateProject } from "./evaluate.js";
+export { queryEvidence, queryGoalDelta } from "./query.js";
+export { checkConfiguration, discoverInputs } from "./doctor.js";
 export {
   compareGoalContracts,
   createGoalDelta,